requirements.txt
numpy
pillow
torch
torchvision
matplotlib
pyyaml
pycuda
tensorrt